move triangle count algos from program algo module to algorithms
