+++ docbrown/db/src/algorithms/mod.rs
@@ -1,6 +1,6 @@
+pub mod connected_components;
 pub mod degree;
 pub mod directed_graph_density;
 pub mod local_clustering_coefficient;
-pub mod triangle_count;
 pub mod reciprocity;
-pub mod connected_components;
+pub mod triangle_count;
