+++ examples/rust/src/bin/hulongbay/main.rs
@@ -13,12 +13,9 @@ use docbrown_core::tgraph::TemporalGraph;
 use docbrown_core::{state, utils};
 use docbrown_core::{Direction, Prop};
 use docbrown_db::algorithms::connected_components::weakly_connected_components;
-use docbrown_db::algorithms::triangle_count::global_triangle_count;
+use docbrown_db::algorithms::triangle_count::{global_triangle_count, triangle_counting_fast};
 use docbrown_db::csv_loader::csv::CsvLoader;
-use docbrown_db::program::algo::triangle_counting_fast;
-use docbrown_db::program::{
-    GlobalEvalState, Program, TriangleCountS1, TriangleCountS2, TriangleCountSlowS2,
-};
+use docbrown_db::program::{GlobalEvalState, Program};
 use itertools::Itertools;
 use regex::Regex;
 use serde::Deserialize;
